+++ stroom-authentication/stroom-authentication-api/src/main/java/stroom/authentication/resources/authentication/v1/ChangePasswordRequest.java
@@ -18,20 +18,14 @@
 
 package stroom.authentication.resources.authentication.v1;
 
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
-
 import javax.validation.constraints.NotNull;
 
-@ApiModel(description = "A request to change a user's password.")
 public class ChangePasswordRequest extends ResetPasswordRequest {
 
     @NotNull
-    @ApiModelProperty(value = "The old password.", required = true)
     private String oldPassword;
 
     @NotNull
-    @ApiModelProperty(value = "The email address of the user we're changing the password for.", required = true)
     private String email;
 
     public String getEmail() {