+++ stroom-authentication/stroom-authentication-impl/src/main/java/stroom/authentication/resources/token/v1/TokenServiceImpl.java
@@ -16,7 +16,7 @@ import javax.ws.rs.BadRequestException;
 import java.time.Instant;
 import java.util.Optional;
 
-public class TokenService {
+public class TokenServiceImpl implements TokenService {
     private final TokenDao dao;
     private final SecurityContext securityContext;
     private StroomConfig stroomConfig;
@@ -26,13 +26,13 @@ public class TokenService {
     private final StroomEventLoggingService stroomEventLoggingService;
 
     @Inject
-    public TokenService(final TokenDao dao,
-                        final SecurityContext securityContext,
-                        final StroomConfig stroomConfig,
-                        final TokenVerifier tokenVerifier,
-                        final stroom.authentication.resources.user.v1.UserService userService,
-                        final stroom.security.impl.UserService securityUserService,
-                        final StroomEventLoggingService stroomEventLoggingService){
+    public TokenServiceImpl(final TokenDao dao,
+                            final SecurityContext securityContext,
+                            final StroomConfig stroomConfig,
+                            final TokenVerifier tokenVerifier,
+                            final stroom.authentication.resources.user.v1.UserService userService,
+                            final stroom.security.impl.UserService securityUserService,
+                            final StroomEventLoggingService stroomEventLoggingService){
 
         this.dao = dao;
         this.securityContext = securityContext;
@@ -44,6 +44,7 @@ public class TokenService {
     }
 
 
+    @Override
     public SearchResponse search(SearchRequest searchRequest) {
         checkPermission();
         // Validate filters
@@ -62,6 +63,7 @@ public class TokenService {
         return results;
     }
 
+    @Override
     public Token create(CreateTokenRequest createTokenRequest){
         checkPermission();
 
@@ -88,36 +90,42 @@ public class TokenService {
         return token;
     }
 
+    @Override
     public void deleteAll() {
         checkPermission();
         dao.deleteAllTokensExceptAdmins();
         stroomEventLoggingService.createAction("DeleteAllApiTokens", "Delete all tokens");
     }
 
+    @Override
     public void delete(int tokenId) {
         checkPermission();
         dao.deleteTokenById(tokenId);
         stroomEventLoggingService.createAction("DeleteApiToken", "Delete a token by ID");
     }
 
+    @Override
     public void delete(String token){
         checkPermission();
         dao.deleteTokenByTokenString(token);
         stroomEventLoggingService.createAction("DeleteApiToken", "Delete a token by the value of the actual token.");
     }
 
+    @Override
     public Optional<Token> read(String token){
         checkPermission();
         stroomEventLoggingService.createAction("ReadApiToken", "Read a token by the string value of the token.");
         return dao.readByToken(token);
     }
 
+    @Override
     public Optional<Token> read(int tokenId) {
         checkPermission();
         stroomEventLoggingService.createAction("ReadApiToken", "Read a token by the token ID.");
         return dao.readById(tokenId);
     }
 
+    @Override
     public void toggleEnabled(int tokenId, boolean isEnabled){
        checkPermission();
         final String userId = securityContext.getUserId();
@@ -129,6 +137,7 @@ public class TokenService {
         }
     }
 
+    @Override
     public Optional<String> verifyToken(String token){
         Optional<Token> tokenRecord = dao.readByToken(token);
         if (!tokenRecord.isPresent()) {
@@ -138,6 +147,7 @@ public class TokenService {
         return usersEmail;
     }
 
+    @Override
     public String getPublicKey() {
         String jwkAsJson = tokenVerifier.getJwk().toJson(JsonWebKey.OutputControlLevel.PUBLIC_ONLY);
         stroomEventLoggingService.createAction("GetPublicApiKey", "Read a token by the token ID.");
@@ -150,7 +160,7 @@ public class TokenService {
         }
     }
 
-    public static Optional<Token.TokenType> getParsedTokenType(String tokenType) {
+    static Optional<Token.TokenType> getParsedTokenType(String tokenType) {
         switch (tokenType.toLowerCase()) {
             case "api":
                 return Optional.of(Token.TokenType.API);
@@ -162,4 +172,6 @@ public class TokenService {
                 return Optional.empty();
         }
     }
+
+
 }