+++ stroom-proxy/stroom-proxy-app/src/main/java/stroom/proxy/app/guice/ProxyModule.java
@@ -14,7 +14,6 @@ import org.slf4j.LoggerFactory;
 import stroom.collection.mock.MockCollectionModule;
 import stroom.dictionary.impl.DictionaryModule;
 import stroom.dictionary.impl.DictionaryStore;
-import stroom.dictionary.impl.NewUiDictionaryResource;
 import stroom.dictionary.impl.NewUiDictionaryResource2;
 import stroom.docstore.api.DocumentResourceHelper;
 import stroom.docstore.api.Serialiser2Factory;
@@ -113,7 +112,6 @@ public class ProxyModule extends AbstractModule {
 
         HealthCheckBinder.create(binder())
                 .bind(ContentSyncService.class)
-                .bind(NewUiDictionaryResource.class)
                 .bind(NewUiDictionaryResource2.class)
                 .bind(FeedStatusResource.class)
                 .bind(ForwardStreamHandlerFactory.class)
@@ -135,7 +133,6 @@ public class ProxyModule extends AbstractModule {
                 .bind(ReceiveDataServlet.class);
 
         GuiceUtil.buildMultiBinder(binder(), RestResource.class)
-                .addBinding(NewUiDictionaryResource.class)
                 .addBinding(NewUiDictionaryResource2.class)
                 .addBinding(ReceiveDataRuleSetResource.class)
                 .addBinding(ReceiveDataRuleSetResourceImpl.class)