+++ CHANGELOG.md
@@ -6,8 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+* Update auth flow for auth-into-stroom integration
 
 ## [v7.0-beta.23] - 2020-03-24
+
 
 * Rename all legacy DB tables to `OLD_`.